move util functions from 1 file into their own file in a utils folder
